Add logic for parsing text with language attributes
This code will walk the tree and find the language attribute of the node.
If the node does not have a specified language attribute, it will search
the parents until it finds a language attribute to inherit from a parent
element.

Example usage:
node index.js http://localhost:8080 --verbose --unicodes -f 200 -l ja